arc: init next

+++ files/board/arpl/overlayfs/opt/arpl/include/addons.sh
@@ -4,16 +4,16 @@
 # 1 - Platform
 # 2 - Kernel Version
 function availableAddons() {
-  while read D; do
-    [ ! -f "${D}/manifest.yml" ] && continue
+  while read -r D; do
+    [[ ! -f ${D}/manifest.yml ]] && continue
     ADDON=$(basename ${D})
     checkAddonExist "${ADDON}" "${1}" "${2}" || continue
     SYSTEM=$(readConfigKey "system" "${D}/manifest.yml")
-    [ "${SYSTEM}" = "true" ] && continue
-    while IFS=': ' read AVAILABLE; do
-    [ "${AVAILABLE}" = "${1}-${2}" ] && ACTIVATE="true" && break || ACTIVATE="false"
+    [[ ${SYSTEM} = true ]] && continue
+    while IFS=': ' read -r AVAILABLE; do
+    [[ ${AVAILABLE} = ${1}-${2} ]] && ACTIVATE="true" && break || ACTIVATE="false"
     done < <(readConfigEntriesArray "available-for" "${D}/manifest.yml")
-    [ "${ACTIVATE}" = "false" ] && continue
+    [[ ${ACTIVATE} = false ]] && continue
     DESC="$(readConfigKey "description" "${D}/manifest.yml")"
     echo -e "${ADDON}\t${DESC}"
   done < <(find "${ADDONS_PATH}" -maxdepth 1 -type d | sort)
@@ -27,11 +27,11 @@ function availableAddons() {
 # Return ERROR if not exists
 function checkAddonExist() {
   # First check generic files
-  if [ -f "${ADDONS_PATH}/${1}/all.tgz" ]; then
+  if [[ -f ${ADDONS_PATH}/${1}/all.tgz ]]; then
     return 0 # OK
   fi
   # Now check specific platform file
-  if [ -f "${ADDONS_PATH}/${1}/${2}-${3}.tgz" ]; then
+  if [[ -f ${ADDONS_PATH}/${1}/${2}-${3}.tgz ]]; then
     return 0 # OK
   fi
   return 1 # ERROR
@@ -42,28 +42,28 @@ function checkAddonExist() {
 # 1 - Addon id
 function installAddon() {
   ADDON="${1}"
-  mkdir -p "${TMP_PATH}/${ADDON}"
+  mkdir -p ${TMP_PATH}/${ADDON}
   HAS_FILES=0
   # First check generic files
-  if [ -f "${ADDONS_PATH}/${ADDON}/all.tgz" ]; then
-    tar -zxf "${ADDONS_PATH}/${ADDON}/all.tgz" -C "${TMP_PATH}/${ADDON}"
+  if [[ -f ${ADDONS_PATH}/${ADDON}/all.tgz ]]; then
+    tar -zxf ${ADDONS_PATH}/${ADDON}/all.tgz -C ${TMP_PATH}/${ADDON}
     HAS_FILES=1
   fi
   # Now check specific platform files
-  if [ -f "${ADDONS_PATH}/${ADDON}/${PLATFORM}-${KVER}.tgz" ]; then
-    tar -zxf "${ADDONS_PATH}/${ADDON}/${PLATFORM}-${KVER}.tgz" -C "${TMP_PATH}/${ADDON}"
+  if [[ -f ${ADDONS_PATH}/${ADDON}/${PLATFORM}-${KVER}.tgz ]]; then
+    tar -zxf ${ADDONS_PATH}/${ADDON}/${PLATFORM}-${KVER}.tgz -C ${TMP_PATH}/${ADDON}
     HAS_FILES=1
   fi
   # Check if addon is available for this platform
-  while IFS=': ' read AVAILABLE; do
-    [ "${AVAILABLE}" = "${PLATFORM}-${KVER}" ] && ACTIVATE="true" && break || ACTIVATE="false"
+  while IFS=': ' read -r AVAILABLE; do
+    [[ ${AVAILABLE} = ${PLATFORM}-${KVER} ]] && ACTIVATE="true" && break || ACTIVATE="false"
   done < <(readConfigEntriesArray "available-for" "${ADDONS_PATH}/${ADDON}/manifest.yml")
   # If has files to copy, copy it, else return error
-  [ ${HAS_FILES} -ne 1 ] || [ ${ACTIVATE} = "false" ] && return 1
-  cp "${TMP_PATH}/${ADDON}/install.sh" "${RAMDISK_PATH}/addons/${ADDON}.sh" 2>"${LOG_FILE}" || dieLog
-  chmod +x "${RAMDISK_PATH}/addons/${ADDON}.sh"
-  [ -d ${TMP_PATH}/${ADDON}/root ] && (cp -R "${TMP_PATH}/${ADDON}/root/"* "${RAMDISK_PATH}/" 2>"${LOG_FILE}" || dieLog)
-  rm -rf "${TMP_PATH}/${ADDON}"
+  [[ ${HAS_FILES} != 1 || ${ACTIVATE} = false ]] && return 1
+  cp -f ${TMP_PATH}/${ADDON}/install.sh ${RAMDISK_PATH}/addons/${ADDON}.sh 2>"${LOG_FILE}" || dieLog
+  chmod +x ${RAMDISK_PATH}/addons/${ADDON}.sh
+  [[ -d ${TMP_PATH}/${ADDON}/root ]] && (cp -R ${TMP_PATH}/${ADDON}/root/* ${RAMDISK_PATH}/ 2>"${LOG_FILE}" || dieLog)
+  rm -rf ${TMP_PATH}/${ADDON:?}
   return 0
 }
 
@@ -72,12 +72,12 @@ function installAddon() {
 # 1 - Addon file path
 # Return name of addon on sucess or empty on error
 function untarAddon() {
-  rm -rf "${TMP_PATH}/addon"
-  mkdir -p "${TMP_PATH}/addon"
-  tar -xaf "${1}" -C "${TMP_PATH}/addon" || return
+  rm -rf ${TMP_PATH}/addon
+  mkdir -p ${TMP_PATH}/addon
+  tar -xaf ${1} -C ${TMP_PATH}/addon || return
   ADDON=$(readConfigKey "name" "${TMP_PATH}/addon/manifest.yml")
-  [ -z "${ADDON}" ] && return
-  rm -rf "${ADDONS_PATH}/${ADDON}"
-  mv "${TMP_PATH}/addon" "${ADDONS_PATH}/${ADDON}"
+  [[ -z ${ADDON} ]] && return
+  rm -rf ${ADDONS_PATH}/${ADDON:?}
+  mv ${TMP_PATH}/addon ${ADDONS_PATH}/${ADDON}
   echo "${ADDON}"
 }
